show absolute battery level

+++ components/services/monitor.h
@@ -21,7 +21,7 @@ extern void (*spkfault_handler_svc)(bool inserted);
 extern bool spkfault_svc(void);
 
 extern float battery_value_svc(void);
-extern uint8_t battery_level_svc(void);
+extern uint16_t battery_level_svc(void);
 
 extern monitor_gpio_t * get_spkfault_gpio(); 
 extern monitor_gpio_t * get_jack_insertion_gpio();
